# [CHORE] processing discovered targets async