--- a/.chloggen/discovering-target-async.yaml
+++ b/.chloggen/discovering-target-async.yaml
@@ -5,7 +5,7 @@ change_type: enhancement
 component: target allocator
 
 # A brief description of the change. Surround your text with quotes ("") if it needs to start with a backtick (`).
-note: processing discovered targets async
+note: Process discovered targets asyncchronously
 
 # One or more tracking issues related to the change
 issues: [1842]