--- a/cmd/otel-allocator/target/discovery.go
+++ b/cmd/otel-allocator/target/discovery.go
@@ -17,6 +17,8 @@ package target
 import (
 	"hash"
 	"hash/fnv"
+	"sync"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/prometheus/client_golang/prometheus"
@@ -37,16 +39,33 @@ var (
 		Name: "opentelemetry_allocator_targets",
 		Help: "Number of targets discovered.",
 	}, []string{"job_name"})
+
+	processTargetsDuration = promauto.NewHistogram(prometheus.HistogramOpts{
+		Name:    "opentelemetry_allocator_process_targets_duration_seconds",
+		Help:    "Duration of processing targets.",
+		Buckets: []float64{1, 5, 10, 30, 60, 120},
+	})
+
+	processTargetGroupsDuration = promauto.NewHistogramVec(prometheus.HistogramOpts{
+		Name:    "opentelemetry_allocator_process_target_groups_duration_seconds",
+		Help:    "Duration of processing target groups.",
+		Buckets: []float64{1, 5, 10, 30, 60, 120},
+	}, []string{"job_name"})
 )
 
 type Discoverer struct {
-	log                  logr.Logger
-	manager              *discovery.Manager
-	close                chan struct{}
-	configsMap           map[allocatorWatcher.EventSource][]*promconfig.ScrapeConfig
-	hook                 discoveryHook
-	scrapeConfigsHash    hash.Hash
-	scrapeConfigsUpdater scrapeConfigsUpdater
+	log                    logr.Logger
+	manager                *discovery.Manager
+	close                  chan struct{}
+	mtxScrape              sync.Mutex // Guards the fields below.
+	configsMap             map[allocatorWatcher.EventSource][]*promconfig.ScrapeConfig
+	hook                   discoveryHook
+	scrapeConfigsHash      hash.Hash
+	scrapeConfigsUpdater   scrapeConfigsUpdater
+	targetSets             map[string][]*targetgroup.Group
+	triggerReload          chan struct{}
+	processTargetsCallBack func(targets map[string]*Item)
+	mtxTargets             sync.Mutex
 }
 
 type discoveryHook interface {
@@ -57,15 +76,17 @@ type scrapeConfigsUpdater interface {
 	UpdateScrapeConfigResponse(map[string]*promconfig.ScrapeConfig) error
 }
 
-func NewDiscoverer(log logr.Logger, manager *discovery.Manager, hook discoveryHook, scrapeConfigsUpdater scrapeConfigsUpdater) *Discoverer {
+func NewDiscoverer(log logr.Logger, manager *discovery.Manager, hook discoveryHook, scrapeConfigsUpdater scrapeConfigsUpdater, setTargets func(targets map[string]*Item)) *Discoverer {
 	return &Discoverer{
-		log:                  log,
-		manager:              manager,
-		close:                make(chan struct{}),
-		configsMap:           make(map[allocatorWatcher.EventSource][]*promconfig.ScrapeConfig),
-		hook:                 hook,
-		scrapeConfigsHash:    nil, // we want the first update to succeed even if the config is empty
-		scrapeConfigsUpdater: scrapeConfigsUpdater,
+		log:                    log,
+		manager:                manager,
+		close:                  make(chan struct{}),
+		triggerReload:          make(chan struct{}, 1),
+		configsMap:             make(map[allocatorWatcher.EventSource][]*promconfig.ScrapeConfig),
+		hook:                   hook,
+		scrapeConfigsHash:      nil, // we want the first update to succeed even if the config is empty
+		scrapeConfigsUpdater:   scrapeConfigsUpdater,
+		processTargetsCallBack: setTargets,
 	}
 }
 
@@ -105,43 +126,110 @@ func (m *Discoverer) ApplyConfig(source allocatorWatcher.EventSource, scrapeConf
 	return m.manager.ApplyConfig(discoveryCfg)
 }
 
-func (m *Discoverer) Watch(fn func(targets map[string]*Item)) error {
-	labelsBuilder := labels.NewBuilder(labels.EmptyLabels())
+func (m *Discoverer) Watch() error {
+	err := m.run(m.manager.SyncCh())
+	if err != nil {
+		m.log.Error(err, "Service Discovery watch event failed")
+		return err
+	}
+	<-m.close
+	m.log.Info("Service Discovery watch event stopped: discovery manager closed")
+	return nil
+}
+
+func (m *Discoverer) UpdateTsets(tsets map[string][]*targetgroup.Group) {
+	m.mtxScrape.Lock()
+	m.targetSets = tsets
+	m.mtxScrape.Unlock()
+}
+
+func (m *Discoverer) reloader() {
+	reloadIntervalDuration := model.Duration(5 * time.Second)
+	ticker := time.NewTicker(time.Duration(reloadIntervalDuration))
+
+	defer ticker.Stop()
+
 	for {
 		select {
 		case <-m.close:
-			m.log.Info("Service Discovery watch event stopped: discovery manager closed")
-			return nil
-		case tsets := <-m.manager.SyncCh():
-			m.ProcessTargets(labelsBuilder, tsets, fn)
+			return
+		case <-ticker.C:
+			select {
+			case <-m.triggerReload:
+				m.Reload()
+			case <-m.close:
+				return
+			}
 		}
 	}
 }
 
-func (m *Discoverer) ProcessTargets(builder *labels.Builder, tsets map[string][]*targetgroup.Group, fn func(targets map[string]*Item)) {
+func (m *Discoverer) Reload() {
+	m.mtxScrape.Lock()
+	var wg sync.WaitGroup
 	targets := map[string]*Item{}
+	timer := prometheus.NewTimer(processTargetsDuration)
+	defer timer.ObserveDuration()
 
-	for jobName, tgs := range tsets {
-		var count float64 = 0
-		for _, tg := range tgs {
-			builder.Reset(labels.EmptyLabels())
-			for ln, lv := range tg.Labels {
+	for jobName, groups := range m.targetSets {
+		wg.Add(1)
+		// Run the sync in parallel as these take a while and at high load can't catch up.
+		go func(jobName string, groups []*targetgroup.Group) {
+			m.processTargetGroups(jobName, groups, targets)
+			wg.Done()
+		}(jobName, groups)
+	}
+	m.mtxScrape.Unlock()
+	wg.Wait()
+	m.processTargetsCallBack(targets)
+}
+
+func (m *Discoverer) processTargetGroups(jobName string, groups []*targetgroup.Group, targets map[string]*Item) {
+	builder := labels.NewBuilder(labels.Labels{})
+	timer := prometheus.NewTimer(processTargetGroupsDuration.WithLabelValues(jobName))
+	defer timer.ObserveDuration()
+	var count float64 = 0
+	for _, tg := range groups {
+		builder.Reset(labels.EmptyLabels())
+		for ln, lv := range tg.Labels {
+			builder.Set(string(ln), string(lv))
+		}
+		groupLabels := builder.Labels()
+		for _, t := range tg.Targets {
+			count++
+			builder.Reset(groupLabels)
+			for ln, lv := range t {
 				builder.Set(string(ln), string(lv))
 			}
-			groupLabels := builder.Labels()
-			for _, t := range tg.Targets {
-				count++
-				builder.Reset(groupLabels)
-				for ln, lv := range t {
-					builder.Set(string(ln), string(lv))
-				}
-				item := NewItem(jobName, string(t[model.AddressLabel]), builder.Labels(), "")
-				targets[item.Hash()] = item
-			}
+			item := NewItem(jobName, string(t[model.AddressLabel]), builder.Labels(), "")
+			m.mtxTargets.Lock()
+			targets[item.Hash()] = item
+			m.mtxTargets.Unlock()
 		}
-		targetsDiscovered.WithLabelValues(jobName).Set(count)
 	}
-	fn(targets)
+	targetsDiscovered.WithLabelValues(jobName).Set(count)
+}
+
+// Run receives and saves target set updates and triggers the scraping loops reloading.
+// Reloading happens in the background so that it doesn't block receiving targets updates.
+func (m *Discoverer) run(tsets <-chan map[string][]*targetgroup.Group) error {
+	go m.reloader()
+	for {
+		select {
+		case ts := <-tsets:
+			m.log.Info("Service Discovery watch event received", "targets groups", len(ts))
+			m.UpdateTsets(ts)
+
+			select {
+			case m.triggerReload <- struct{}{}:
+			default:
+			}
+
+		case <-m.close:
+			m.log.Info("Service Discovery watch event stopped: discovery manager closed")
+			return nil
+		}
+	}
 }
 
 func (m *Discoverer) Close() {