**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This change addresses issue [#3512](https://github.com/open-telemetry/opentelemetry-operator/issues/3512). The implementation is inspired by the Prometheus Scrape Manager, as referenced in the source code here: [Prometheus Scrape Manager](https://github.com/prometheus/prometheus/blob/af2a1cb10c89de496cd4309ac532624b34112d74/scrape/manager.go#L121).

Below are the `benchstat` results comparing the current `main` branch with the proposed changes in this PR. These results highlight the performance improvements and trade-offs introduced by this update.
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator
cpu: Apple M1 Pro
                                                      │ bench_targets-serial.txt │     bench_targets-concurrent.txt     │
                                                      │          sec/op          │    sec/op     vs base                │
ProcessTargets/consistent-hashing-10                                 3.658 ±  5%    1.087 ±  9%  -70.30% (p=0.000 n=10)
ProcessTargets/per-node-10                                         3586.9m ±  3%   995.7m ± 14%  -72.24% (p=0.000 n=10)
ProcessTargets/least-weighted-10                                   3627.1m ±  3%   985.5m ±  7%  -72.83% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node-10                        1007.7m ± 11%   959.8m ±  8%        ~ (p=0.529 n=10)
ProcessTargetsWithRelabelConfig/least-weighted-10                   994.0m ±  8%   977.5m ± 10%        ~ (p=0.481 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing-10              1010.2m ± 10%   927.3m ±  2%   -8.21% (p=0.002 n=10)
geomean                                                              1.907         987.6m        -48.23%

                                                      │ bench_targets-serial.txt │    bench_targets-concurrent.txt     │
                                                      │           B/op           │     B/op      vs base               │
ProcessTargets/consistent-hashing-10                                3.158Gi ± 0%   3.178Gi ± 0%  +0.64% (p=0.000 n=10)
ProcessTargets/per-node-10                                          3.158Gi ± 0%   3.178Gi ± 0%  +0.64% (p=0.000 n=10)
ProcessTargets/least-weighted-10                                    3.158Gi ± 0%   3.178Gi ± 0%  +0.64% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node-10                         3.180Gi ± 0%   3.180Gi ± 0%       ~ (p=0.190 n=10)
ProcessTargetsWithRelabelConfig/least-weighted-10                   3.180Gi ± 0%   3.180Gi ± 0%       ~ (p=0.529 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing-10               3.180Gi ± 0%   3.180Gi ± 0%       ~ (p=0.684 n=10)
geomean                                                             3.169Gi        3.179Gi       +0.32%

                                                      │ bench_targets-serial.txt │    bench_targets-concurrent.txt    │
                                                      │        allocs/op         │  allocs/op   vs base               │
ProcessTargets/consistent-hashing-10                                 4.161M ± 0%   4.221M ± 0%  +1.45% (p=0.002 n=10)
ProcessTargets/per-node-10                                           4.161M ± 0%   4.221M ± 0%  +1.45% (p=0.000 n=10)
ProcessTargets/least-weighted-10                                     4.161M ± 0%   4.220M ± 0%  +1.43% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node-10                          4.236M ± 0%   4.237M ± 0%       ~ (p=0.724 n=10)
ProcessTargetsWithRelabelConfig/least-weighted-10                    4.237M ± 0%   4.237M ± 0%       ~ (p=0.631 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing-10                4.236M ± 0%   4.237M ± 0%       ~ (p=0.436 n=10)
geomean                                                              4.198M        4.229M       +0.72%
```
**Link to tracking Issue(s):** <Issue number if applicable>

- Resolves: #3512

**Testing:** Run Unit Tests and BenchMarch Tests

**Documentation:** N/A