@nicolastakashi 2024-12-17
```goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator
cpu: Apple M1 Pro
BenchmarkProcessTargets/per-node-10            	       1	1122291333 ns/op	3431733448 B/op	 4417034 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 985838084 ns/op	3414612288 B/op	 4234554 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 968455875 ns/op	3414472688 B/op	 4234142 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 902708542 ns/op	3414559424 B/op	 4235042 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 914759875 ns/op	3414536968 B/op	 4234955 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 861704500 ns/op	3414609664 B/op	 4235733 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 824441770 ns/op	3414526040 B/op	 4234824 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	1053985854 ns/op	3414310288 B/op	 4232579 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	1014306334 ns/op	3414348072 B/op	 4232986 allocs/op
BenchmarkProcessTargets/per-node-10            	       2	 927938792 ns/op	3414307512 B/op	 4232577 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 838161250 ns/op	3414583368 B/op	 4235406 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 950306875 ns/op	3414322416 B/op	 4232724 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 892734354 ns/op	3414382264 B/op	 4233430 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 882699542 ns/op	3414367800 B/op	 4233267 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 918699980 ns/op	3414305280 B/op	 4232622 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 827566146 ns/op	3414596352 B/op	 4234735 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 932759167 ns/op	3414521768 B/op	 4234829 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 835300375 ns/op	3414518352 B/op	 4234696 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 862945354 ns/op	3414511400 B/op	 4234795 allocs/op
BenchmarkProcessTargets/least-weighted-10      	       2	 917244250 ns/op	3414306992 B/op	 4232574 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 908266292 ns/op	3414353496 B/op	 4233097 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 872114500 ns/op	3414446104 B/op	 4233992 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 863686250 ns/op	3414349784 B/op	 4233032 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 884526562 ns/op	3414359912 B/op	 4233131 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 872686104 ns/op	3414537528 B/op	 4234973 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 899104500 ns/op	3414298424 B/op	 4232552 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 856262916 ns/op	3414461864 B/op	 4234234 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 887320250 ns/op	3414394752 B/op	 4233453 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 910004980 ns/op	3414333144 B/op	 4232886 allocs/op
BenchmarkProcessTargets/consistent-hashing-10  	       2	 880362000 ns/op	3414589832 B/op	 4235591 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 894908562 ns/op	3415972280 B/op	 4250871 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 884152000 ns/op	3415784728 B/op	 4248901 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 880393438 ns/op	3415879584 B/op	 4249947 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	1006655375 ns/op	3415878896 B/op	 4249814 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 875715896 ns/op	3415914608 B/op	 4250269 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 912449271 ns/op	3415745264 B/op	 4248512 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 881933730 ns/op	3415738312 B/op	 4248428 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 851389396 ns/op	3415938952 B/op	 4250570 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 860880292 ns/op	3415879320 B/op	 4249898 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/least-weighted-10         	       2	 910038542 ns/op	3415783504 B/op	 4248931 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 846392438 ns/op	3415940752 B/op	 4250580 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 922212542 ns/op	3415764984 B/op	 4248759 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 893194479 ns/op	3415929848 B/op	 4250460 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 888838542 ns/op	3415776144 B/op	 4248804 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 876409104 ns/op	3415805960 B/op	 4249196 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 924206646 ns/op	3415770592 B/op	 4248793 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 850870812 ns/op	3415931992 B/op	 4250439 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 884762875 ns/op	3415814584 B/op	 4249200 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 896129083 ns/op	3415959216 B/op	 4250795 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/consistent-hashing-10     	       2	 900787250 ns/op	3415772504 B/op	 4248838 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 930492312 ns/op	3415914392 B/op	 4250305 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 861972688 ns/op	3415985264 B/op	 4250985 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 872793875 ns/op	3415783328 B/op	 4248945 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 889231812 ns/op	3415923628 B/op	 4250333 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 875701146 ns/op	3415837688 B/op	 4249438 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 880286062 ns/op	3415972176 B/op	 4250887 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 879769188 ns/op	3415780680 B/op	 4248888 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	1264806020 ns/op	3415740260 B/op	 4248454 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       2	 993177812 ns/op	3415891352 B/op	 4250057 allocs/op
BenchmarkProcessTargetsWithRelabelConfig/per-node-10               	       1	1084294000 ns/op	3416297216 B/op	 4249305 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator	177.920s
```

Bench results after lock per job